backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
